% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsspiral.R
\name{detect_tsspiral_frequency}
\alias{detect_tsspiral_frequency}
\title{Detect Time-Series Frequency}
\usage{
detect_tsspiral_frequency(x)
}
\arguments{
\item{x}{Date vector.}
}
\value{
A character string describing the frequency.
}
\description{
Internal function used to detect the approximate frequency of
the supplied time series.
}
\keyword{internal}
