% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsspiral.R
\name{expand_tsspiral_daily}
\alias{expand_tsspiral_daily}
\title{Expand Time Series to Daily Resolution}
\usage{
expand_tsspiral_daily(data, frequency)
}
\arguments{
\item{data}{Data frame containing \code{x}.}

\item{frequency}{Detected frequency.}
}
\value{
A data frame with daily observations.
}
\description{
Internal function that converts weekly, monthly, quarterly,
and annual observations into daily observations.
}
\keyword{internal}
