% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsspiral.R
\docType{data}
\name{GeomTSSpiral}
\alias{GeomTSSpiral}
\title{Spiral Time-Series Geometry}
\format{
An object of class \code{GeomTSSpiral} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTSSpiral
}
\description{
Internal ggproto geometry used by \code{\link[=geom_tsspiral]{geom_tsspiral()}}.
}
\keyword{internal}
