% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsspiralbar.R
\docType{data}
\name{GeomTSSpiralBar}
\alias{GeomTSSpiralBar}
\title{Spiral Time-Series Bar Geometry}
\format{
An object of class \code{GeomTSSpiralBar} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTSSpiralBar
}
\description{
Internal ggproto geometry used by \code{\link[=geom_tsspiralbar]{geom_tsspiralbar()}}.
}
\keyword{internal}
