% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsspiralmiss.R
\docType{data}
\name{GeomTSSpiralMiss}
\alias{GeomTSSpiralMiss}
\title{Spiral Missingness Geometry}
\format{
An object of class \code{GeomTSSpiralMiss} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTSSpiralMiss
}
\description{
Internal ggproto geometry used by \code{\link[=geom_tsspiralmiss]{geom_tsspiralmiss()}}.
}
\keyword{internal}
