% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsspiral.R
\docType{data}
\name{StatTSSpiral}
\alias{StatTSSpiral}
\title{Statistical Transformation for Spiral Time Series}
\format{
An object of class \code{StatTSSpiral} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatTSSpiral
}
\description{
Statistical Transformation for Spiral Time Series
}
\keyword{internal}
