% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsspiralbar.R
\docType{data}
\name{StatTSSpiralBar}
\alias{StatTSSpiralBar}
\title{Statistical Transformation for Spiral Bars}
\format{
An object of class \code{StatTSSpiralBar} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatTSSpiralBar
}
\description{
Statistical Transformation for Spiral Bars
}
\keyword{internal}
