% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsspiralmiss.R
\docType{data}
\name{StatTSSpiralMiss}
\alias{StatTSSpiralMiss}
\title{Statistical Transformation for Spiral Missingness}
\format{
An object of class \code{StatTSSpiralMiss} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatTSSpiralMiss
}
\description{
Internal statistical transformation used by
\code{\link[=geom_tsspiralmiss]{geom_tsspiralmiss()}}.
}
\keyword{internal}
