% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsspiral.R
\name{tsspiral_calendar_labels}
\alias{tsspiral_calendar_labels}
\title{Generate Spiral Calendar Labels}
\usage{
tsspiral_calendar_labels(frequency)
}
\arguments{
\item{frequency}{Detected time-series frequency.}
}
\value{
Data frame containing label positions and labels.
}
\description{
Internal function generating labels for the outer cycle.
}
\keyword{internal}
