% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsspiral.R
\name{tsspiral_day}
\alias{tsspiral_day}
\title{Convert Date to Fixed 365-Day Calendar}
\usage{
tsspiral_day(x)
}
\arguments{
\item{x}{Date vector.}
}
\value{
Integer vector from 1 to 365.
}
\description{
Internal function that maps dates onto a fixed 365-day calendar.
}
\keyword{internal}
